Sequence changes to max-age and forwarding-delay
The mstp daemon only gets to see 1 value at a time, but at all times
we need to ensure 2*(forwarding-delay-1) >= max-age.

In order to not upset mstpd, figure out which attribute to set
first. Check the constraint using the YANG value of one of the
attributes and the current mstpd value of the other attribute and
update accordingly.

Fixes: VRVDR-49427